chore: the purge continues
