+++ .github/workflows/ci.yml
@@ -62,6 +62,10 @@ jobs:
       - name: 🎁 dotnet tool restore
         run: |
           dotnet tool restore
+      - name: Workloads
+        id: workloads
+        run: |
+          dotnet nuke Workloads --skip
       - name: 🎁 Restore
         id: restore
         run: |
@@ -78,6 +82,10 @@ jobs:
         id: pack
         run: |
           dotnet nuke Pack --skip
+      - name: Git Hub Actions
+        id: gitHubActions
+        run: |
+          dotnet nuke GitHubActions --skip
       - name: 🐿 Publish Coverage
         uses: codecov/codecov-action@v3.1.4
         with:
